backend/requirements.txt
Django==5.2.12
djangorestframework==3.17.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.2
python-dotenv==1.2.2
